refactor: 접속과 관련된 패키지 생성
