+++ src/test/java/com/gnimty/communityapiserver/domain/chat/service/ChatServiceTest.java
@@ -15,16 +15,13 @@ import com.gnimty.communityapiserver.domain.chat.repository.Chat.ChatRepository;
 import com.gnimty.communityapiserver.domain.chat.repository.ChatRoom.ChatRoomRepository;
 import com.gnimty.communityapiserver.domain.chat.repository.User.UserRepository;
 import com.gnimty.communityapiserver.domain.chat.service.dto.UserWithBlockDto;
-import com.gnimty.communityapiserver.global.auth.OnlineInChatRoomManager;
-import com.gnimty.communityapiserver.global.auth.OnlineInChatRoomManager.UserCurrentlyInChatRoom;
+import com.gnimty.communityapiserver.global.connect.OnlineInChatRoomManager;
 import com.gnimty.communityapiserver.global.constant.MessageRequestType;
 import com.gnimty.communityapiserver.global.constant.Status;
 import com.gnimty.communityapiserver.global.constant.Tier;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
